taptenc/timed-automata: add datatypes to store timed automata models
The datatypes are oriented at the uppaal functionalities of timed
automata systems, but are simplified to our expected needs. At the
moment constraints are just strings, a proper model may be future work.
Visual information such as node positions are not contained as we
distinguish between theoretic model and visualization.