printer: add xta printer
basic draft for an xta printer. UGI printing is not available yet and
urgent locations are just being ignored. There is also a general problem
because chars like < > and & are currently written to be compatible with
xml and therefore encoded. This is wrong in the xta format. Currently
one has to manually take care of this.