taptenc/encoder: move filter utils to own class
Applying and modifying filters can be encapsulated into a spearate
class. Also the whole procedure to keep track of base states from copies
is refactored. Previously the map base_ids explicitly kept track on the
relations between copied states. To simplify this, we can encode the
relations in the state prefixes directly.