taptenc: diferentiate between unary, binary and chain encoding info
To represent the activaton rules given by interconnection constraints we
essentially have to match plan actions to the necessary encoding info
(EncICInfo). However the structure of the encoding info depends on the type of
constraints, in particular we differentiate between unary, binary and chain
constraints. To downcast from EncICInfo to the respective subclass we
have to actually hold pointers to the EncICInfo, else slicing occurs.