direct-encoder: prune TAs without outgoing transitions
prune TAs without outgoing transitions because they cannot be visited on
any run to the query state. Currently this is only done implicitly upon
merging the final system together. This is not optimal as 'dead' TA
copies are carried through the encoding and cause unecessary overhead.
However, compared to the model checking procedure the encoding generaton
time is probably neglectable anyways.