taptenc/filter: correctly handle state ids being suffixes of others
There was a problem when filtering states based on prefixes and
suffixes. In particular, when the suffix after BASE_SEP was filter and
one state id was a suffix of another states id. E.g. a state named
calibrated and one named uncalibrated. filtering for calibrated could
match the state uncalibrated. To prevent this, assume that if the string
to filter does not contain BASE_SEP anymore it should exactly match the
suffix.