utap-trace-parser: make matching with base transitions robust
To ensure base transitions can be matched when reading a taken
transition, piecewise check every atomic constraints of possible
candidate transition for an occurance in the parsed transition. If all
constraints are contained, then the correct base transition is found.
This prevents mismatches due to reordered constraints as they may occur
when merging systems.