taptenc/filter: add inverse filtering
The default filter interprets the filter vector as basis of states where
effects should be applied to and ignores states not contained in the
filter vector. So everything that 'runs through' the filter is not
considered while everything that is 'caught' by the filter is. One might
want to invert this behavior by applying effects to states that 'run
through' the filter and ignoring states that are 'caught'.

To realize this inverse filtering it suffices to invert all checks
of search queries of the filter vector.