xml-printer: move private functions to local namespace
The helper functions for the xml printer do not need to be part of the
XMLPrinter class. In fact the only reasone for the XMLPrinter to be a
class and not a namespace is, that a unified printing interface should
be provided via inheritance.

The benefit of hiding the helper functions from the class is, that the
header becomes more readable.