taptenc/encoder-utils: add :: to std parameters
Doxygen cannot match functions form the header files that are defined
within the taptenc namespace (and therefore have to escape the namespace
before using functionalities from the std:: namespace via ::std::) with
those in the source file, where the toplevel namespace is not taptenc
(because we are utilizing the 'using' directive instead). To fix that,
we can add the (unecessary) namespace escape :: before each function
parameter.