taptenc: remove noisy debug output
Occasions where a clock is already present when adding a clock to an
update set can be intentional (e.g. when joining update sets of
transitions in a product construction).

Similarly it may occur that queries for uppercase markers fail during
encoding (because the default states are not altered in parts of the
system).