taptenc/utap-trace-parser: calculate max delay in symbolic state
The maximal delay can be determined by finding the first upper bound to
be crossed once time elapses assuming all clocks are on their lowest
possible value.

This also means that the max delay is only viable in a concrete state,
if all clocks are indeed at their lowest possible value.

To use this effectively one has to adapt the symbolic state to contain
the actual clock values.