encoder: properly apply successor transitions
If constraints concerned with the future are considered, then successor
transitions are used to reach certain states from a target set.
If constraints concerned with the past are considered, then successor
transitions are used to leave certain states from a target set.
The former case is concerned with filtering transition destinations, the
latter has to filter transition sources.